add rich directory manipulation and size functions

+++ etc_backend_fs.go
@@ -6,9 +6,15 @@ import (
 	"os"
 )
 
-func FileController(path string) (*Buffer, error) {
+func FileController(path string, readOnly ...bool) (*Buffer, error) {
 	b := NewBuffer()
-	fb, err := FsBackend(path)
+	ro := false
+
+	if len(readOnly) > 0 && readOnly[0] == true {
+		ro = true
+	}
+
+	fb, err := FsBackend(path, ro)
 	if err != nil {
 		return nil, err
 	}
@@ -18,10 +24,11 @@ func FileController(path string) (*Buffer, error) {
 }
 
 type fsBackend struct {
-	path string
-	file *os.File
-	rpos int64
-	wpos int64
+	readOnly bool
+	path     string
+	file     *os.File
+	rpos     int64
+	wpos     int64
 }
 
 func (f *fsBackend) Flush() error {
@@ -82,6 +89,10 @@ func (f *fsBackend) Read(b []byte) (int, error) {
 }
 
 func (f *fsBackend) Write(b []byte) (int, error) {
+	if f.readOnly {
+		return 0, fmt.Errorf("etc: cannot write to readonly file buffer")
+	}
+
 	f.file.Seek(f.wpos, 0)
 
 	i, err := f.file.Write(b)
@@ -89,7 +100,6 @@ func (f *fsBackend) Write(b []byte) (int, error) {
 		f.wpos += int64(i)
 		return i, nil
 	} else {
-		fmt.Println("ERR", "=>", err)
 		return i, err
 	}
 }
@@ -98,18 +108,28 @@ func (f *fsBackend) Close() error {
 	return f.file.Close()
 }
 
-func FsBackend(path string) (Backend, error) {
+func FsBackend(path string, readOnly bool) (Backend, error) {
 	var err error
 	f := &fsBackend{}
+	f.readOnly = readOnly
 	f.path = path
 
 	if _, err := os.Stat(path); err != nil {
-		_, err2 := os.Create(path)
-		if err2 != nil {
-			return nil, err2
+		if f.readOnly == false {
+			_, err2 := os.Create(path)
+			if err2 != nil {
+				return nil, err2
+			}
+		} else {
+			return nil, fmt.Errorf("etc: could not open file backend: '%s'", err.Error())
 		}
 	}
 
-	f.file, err = os.OpenFile(path, os.O_RDWR, 0700)
+	if f.readOnly {
+		f.file, err = os.OpenFile(path, os.O_RDONLY, 0700)
+	} else {
+		f.file, err = os.OpenFile(path, os.O_RDWR, 0700)
+	}
+
 	return f, err
 }
